Add medium size avatars for use on the user's own settings page.
This adds a medium (500px) size avatar thumbnail, that can be
referenced as `{name}-medium.png`.  It is intended to be used on the
user's own settings page, though we may come up with other users for
high-resolution avatars in the future.

This will automatically generate and upload the medium avatar images
both when a new avatar original is uploaded, or when a medium avatar
is requested (if it hasn't been generated yet).

Note that this implementation does not provide an endpoint for
fetching the medium-size avatar for another user.

[substantially modified by tabbott]