casper: Attempt to fix flaky password change test.
The test just after the password change test has been failing, we
think because in some race condition, the user ends up logged out
after that test (which is the Django behavior, at least).

It's not clear this is the ideal fix, but it should result in the test
passing reliably.