CSS: Transition 'app.css' to SCSS.
This commit transitions all styles under the app.css rule in the
Django pipeline to being compiled by webpack under an app-styles
bundle. It also moves all associated css files to their corresponding
scss files in static/styles/scss.  It removes references to the old
css files from settings.py. It adds the include for the new webpack
bundle in the index.html template and removes the reference to the old
one.