update-prod-static: Don't copy/minify templates and styles.
This is a performance optimization: Rather than copying these files
into the `prod-static` directory and then deleting them, we just don't
copy them over in the first place.

We probably can do the same thing with the JS directory, but I'm
holding off on that for now for more testing.