UserMessage bigint id migration: Live migration part.
How to migrate:
1. Restart server, applying migration 0238 to add the bigint_id column.
2. While the server is running, apply migration 0239, which will handle
copying the id column over to bigint_id for old and new UserMessages,
and then CONCURRENTLY create an index on bigint_id - all this while the
server remains operational.
3. Use the BIGINT_verify_id_equal_bigint_id management command to make
sure things are okay (this may require stopping the server, for
all the UserMessage objects to be in-sync with their database rows).