pointer: Remove have_initial_messages code.
This was implemented in 2012 to avoid showing a loading indicator for
fetching messages for users with no message history.  However, the
Zulip onboarding UI always creates some message history, and fetching
history is fast, so this is likely clutter more than a useful
optimization.

This change results in the frontend tests for ui_init trying to do a
fetch, which is a unit testing mess.  Perhaps the direction I've been
attempting should be discarded and replaced with another approach?